# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "reyhanyanz@gamil.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "login a user",
                "description": "login a user",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "register a user",
                "description": "register a user",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/auth/password/email": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "send email to reset password a user",
                "description": "send email to reset password a user",
                "operationId": "sendPasswordResetLinkEmail",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendEmailRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/auth/password/reset": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "reset password a user",
                "description": "reset password a user",
                "operationId": "resetPasswordAUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LoginRequest": {
                "title": "Login request",
                "description": "Login request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email of an existing user",
                        "type": "string",
                        "example": "zahra@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "password of user",
                        "type": "string",
                        "example": "12345678"
                    }
                },
                "type": "object"
            },
...
            "RegisterRequest": {
                "title": "Register request",
                "description": "Register request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name of an user",
                        "type": "string",
                        "example": "zahra"
                    },
                    "email": {
                        "title": "email",
                        "description": "email of an user",
                        "type": "string",
                        "example": "zahra@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "password of user",
                        "type": "string",
                        "example": "12345678"
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "confirmation of password",
                        "type": "string",
                        "example": "12345678"
                    }
                },
                "type": "object"
            },
            "ResetRequest": {
                "title": "Reset request",
                "description": "Reset request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email of an user",
                        "type": "string",
                        "example": "zahra@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "password of user",
                        "type": "string",
                        "example": "12345678"
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "confirmation of password",
                        "type": "string",
                        "example": "12345678"
                    },
                    "token": {
                        "title": "token",
                        "description": "token that is send to email and it has limited life time",
                        "type": "string",
                        "example": "7a246c3fdd83f204af6f80544bfeebc614aed8cfaf81e31221c33775ce05c26f"
                    }
                },
                "type": "object"
            },
...
            "SendEmailRequest": {
                "title": "SendEmail request",
                "description": "SendEmail request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email of an user",
                        "type": "string",
                        "example": "zahra@gmail.com"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}